requirements.txt
requests
prometheus-client
falcon
argparse
pyyaml
pyOpenSSL